:recycle: MOD: extract vue component for streak
